fix(BirdID): stop shipping the deleted endpoint's client and buffering a copy
ai-inference.ts still POSTed to /api/identify-bird, and App.tsx prefetched it,
so every user downloaded a chunk whose only network call 404s. Its two types
were all that remained in use, so they move to the adapter and the module and
its fixture replay suite go. The suite only exercised the removed endpoint,
which is why the test count drops by 143.

preloadAssets cloned every response to read progress from, but only consumed
the clone when a progress callback was passed. The engine's own load passes
none, so the browser buffered a second full copy of each asset, up to 24 MiB
for the weights.

Cache keys carry the model version while the cache name does not, so a bump
stranded the previous 62 MiB forever. Prune anything outside the current set.
